+++ packages/terra-compact-interactive-list/CHANGELOG.md
@@ -2,6 +2,9 @@
 
 ## Unreleased
 
+* Changed
+  * Keyboard navigation updated to wrap to the next/previous line at the end/start of a visual row.
+
 ## 1.20.0 - (April 23, 2024)
 
 * Changed